+++ tests/testthat/test-odbc-mysql.R
@@ -1,11 +1,20 @@
 context("odbc-mysql")
 
-skip("odbc")
+skip_on_cran()
 
-db <- dbxConnect(adapter=odbc::odbc(), driver="/usr/local/lib/libmaodbc.dylib", database="dbx_test")
+if (isMac()) {
+  # brew install mariadb-connector-odbc
+  driver <- "/opt/homebrew/lib/mariadb/libmaodbc.dylib"
+} else {
+  # apt-get install odbc-mariadb
+  driver <- "/usr/lib/x86_64-linux-gnu/odbc/libmaodbc.so"
+}
+
+skip_if_not(file.exists(driver))
+
+db <- dbxConnect(adapter=odbc::odbc(), driver=driver, database="dbx_test")
 
 dbxExecute(db, "DROP TABLE IF EXISTS events")
-json_type <- if (Sys.getenv("TRAVIS") == "") "JSON" else "TEXT"
-dbxExecute(db, paste0("CREATE TABLE events (id INT NOT NULL AUTO_INCREMENT PRIMARY KEY, city VARCHAR(255), counter INT, bigcounter BIGINT, speed FLOAT, distance DECIMAL(20, 18), created_on DATE, updated_at DATETIME(6), deleted_at TIMESTAMP(6) NULL DEFAULT NULL, open_time TIME, active BOOLEAN, properties ", json_type, ", image BLOB)"))
+dbxExecute(db, paste0("CREATE TABLE events (id INT NOT NULL AUTO_INCREMENT PRIMARY KEY, city VARCHAR(255), counter INT, bigcounter BIGINT, speed FLOAT, distance DECIMAL(20, 18), created_on DATE, updated_at DATETIME(6), deleted_at TIMESTAMP(6) NULL DEFAULT NULL, open_time TIME(3), active BOOLEAN, properties JSON, image BLOB)"))
 
 runTests(db)