+++ NEWS.md
@@ -1,6 +1,12 @@
+## 0.4.0 (2025-03-17)
+
+- Added support for DuckDB
+- Added support for `returning` option for `dbxInsert` and `dbxUpsert` for SQLite 3.35.0+ and SQL Server
+- Fixed typecasting with ODBC for MariaDB
+
 ## 0.3.2 (2024-06-02)
 
-- Added support for `returning` option for `dbxUpsert` for MariaDB 10.5+
+- Added support for `returning` option for `dbxInsert` and `dbxUpsert` for MariaDB 10.5+
 
 ## 0.3.1 (2023-12-11)